Abstract out the common elements of the popup classes into one class
Ticket: https://jira.csc.calpoly.edu/browse/LUC7-142
