+++ __tests__/popup.test.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import {mount, shallow} from 'enzyme'
-import PopupModal from "../src/PopupModal.jsx";
+import ItemView from "../src/ItemView.jsx";
 
 import { configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
@@ -10,11 +10,11 @@ configure({ adapter: new Adapter() });
 
 
 
-describe('PopupModal', () => {
+describe('ItemView', () => {
   it('is open', () => {
     const exampleData = {fields : 'null', classifications : 'null'}
 	const wrapper = shallow(
-		<PopupModal
+		<ItemView
 		  onClose={null}
 		  open={true}
 		  modalData={exampleData}
